Delete JMXFetcher interface
